+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 2.1.0 (?)
 Changes in this release:
+- Fix bug where the `project` fixture doesn't reset the current working directory in the cleanup stage.
 
 # v 2.0.0 (2022-01-24)
 Changes in this release: